refactor: file names changed from CapFirst to hyphens
