+++ src/core/components/canvas/canvas-area.tsx
@@ -1,6 +1,6 @@
-import { FallbackError } from "@/core/components/FallbackError";
 import { Breadcrumb } from "@/core/components/canvas/Breadcrumb";
 import StaticCanvas from "@/core/components/canvas/static/static-canvas";
+import { FallbackError } from "@/core/components/fallback-error";
 import { useBuilderProp, useCodeEditor } from "@/core/hooks";
 import { Skeleton } from "@/ui/shadcn/components/ui/skeleton";
 import { noop } from "lodash-es";
