+++ src/core/components/hot-keys.tsx
