+++ src/core/components/settings/settings-panel.tsx
@@ -1,4 +1,4 @@
-import { FallbackError } from "@/core/components/FallbackError";
+import { FallbackError } from "@/core/components/fallback-error";
 import BlockSettings from "@/core/components/settings/block-settings";
 import BlockStyling from "@/core/components/settings/BlockStyling";
 import { BlockAttributesEditor } from "@/core/components/settings/new-panel/BlockAttributesEditor";
